feat(app): integrate local S2S with Qwen3-Omni and improve handler logic
feat(moves): ensure head poses are consistently cast to float32

fix(openai_realtime): refactor shutdown to handle async cleanup properly

fix(tools): add type checks and ensure compatibility with GotoQueueMove

docs: add "whats-next" for local S2S integration and future optimizations

+++ src/reachy_mini_conversation_app/dance_emotion_moves.py
@@ -125,12 +125,12 @@ class GotoQueueMove(Move):  # type: ignore
 
             # Use start pose if available, otherwise neutral
             if self.start_head_pose is not None:
-                start_pose = self.start_head_pose
+                start_pose = self.start_head_pose.astype(np.float64)
             else:
                 start_pose = create_head_pose(0, 0, 0, 0, 0, 0, degrees=True)
 
             # Interpolate head pose
-            head_pose = linear_pose_interpolation(start_pose, self.target_head_pose, t_clamped)
+            head_pose = linear_pose_interpolation(start_pose, self.target_head_pose.astype(np.float64), t_clamped)
 
             # Interpolate antennas - return as numpy array
             antennas = np.array(
